pyproject.toml
[project]
name = "auto-gmail-uploader"
version = "0.1.0"
description = "Automatically upload Gmail newsletters to Google Drive in Markdown format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "markdownify>=1.2.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
